Abstract item out of RssParser
