+++ rss/parser/__init__.py
@@ -1,0 +1,3 @@
+from rss.parser.document import RssParser
+
+__all__ = [RssParser]
